requirements.txt
# python 3.8.3
jsonschema==3.2.0
pytest==6.1.2
PyYAML==5.3.1
requests==2.25.0
selenium==3.141.0
